server/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-dotenv
google-genai
reportlab
